refactor: modify metadater
